Refactor: oauth related code
